feat: adds new optional Developer Kiosk with `kiosk.json`
This can easily used with built in auth checker.

The `kiosk.json` file is a configuration file used to manage developer 
access to the kiosk system. Each developer's unique details, roles, 
permissions, and active status are stored under their username, 
providing a structured and organized format for managing access. This 
file is located in the `configs/` directory and serves as the primary 
reference for access control.