feat: adds checks for `kiosk` subdomain like: `kiosk.example.com`
This update introduces functionality to check for incoming requests 
targeting the `kiosk` subdomain, such as 
`{kiosk-domain-id}.example.com`. The `kiosk-domain-id` can be configured 
through either the `kiosk.json` file or the `.env` file. When both are 
present, the value of `KIOSK_DOMAIN_ID` specified in `.env` will take 
precedence over the `kiosk.panel.id` value in `kiosk.json`.

Additionally, for the `kiosk` functionality to be active, the 
`panel.enabled` value must also be set to `true`. If `panel.enabled` is 
not enabled, the middleware will skip the `kiosk` checks and proceed 
with the subsequent middlewares as normal. This ensures that `kiosk` 
operations are only executed when explicitly configured and enabled.